% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{cgd}
\alias{cgd}
\title{Obtain an approximate Newton direction}
\usage{
cgd(g, H, wk)
}
\arguments{
\item{g}{the first-order gradient of the objective function}

\item{H}{the second-order gradient of the objective function or we can say Hessian matrix}

\item{wk}{the working set, which all but a working set of the matrix entries will not be fixed to zero}
}
\value{
an approximate Newton direction
}
\description{
Using a conjugate gradient method to compute approxiamte Newton direction.
}
\examples{
\dontrun{
x= matrix(c(2,1,1,1,3,2,1,2,4),3)
cgd(f_gra1(x,1,diag(3)),f_gra2(x,1,diag(3)),c(1,5,6,8,9))
}
}
