% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{conjugate}
\alias{conjugate}
\title{A Conjugate Gradient method using Rcpp}
\usage{
conjugate(A, b)
}
\arguments{
\item{A}{the coefficient matrix}

\item{b}{the augmented matrix}
}
\value{
x0 the solution
}
\description{
Solve a system of linear equations Ax=b
}
\examples{
\dontrun{
A = matrix(c(4,-2,-1,-2,4,-2,-1,-2,3),3)
b = c(0,-2,3)
conjugate(A,b)
}
}
