% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{f}
\alias{f}
\title{The objective function we want to minimize}
\usage{
f(x, lambda, s)
}
\arguments{
\item{x}{a positive definite symmetric matrix}

\item{lambda}{a regularization parameter}

\item{s}{the sample covariance}
}
\value{
d value of the objective function
}
\description{
The objective function we want to minimize
}
\examples{
\dontrun{
f(diag(3),1,diag(3))
}
}
