% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{f_gra2c}
\alias{f_gra2c}
\title{Compute second-order gradient}
\usage{
f_gra2c(x, lambda, s)
}
\arguments{
\item{x}{a positive definite symmetric matrix}

\item{lambda}{a regularization parameter}

\item{s}{the sample covariance}
}
\value{
d the second-order gradient
}
\description{
Compute second-order gradient of the objective function
}
\examples{
\dontrun{
f_gra2c(diag(3),1,diag(3))
}
}
