% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{HANA}
\alias{HANA}
\title{Homotopy Approximate Newton Algorithm}
\usage{
HANA(
  n,
  kkk,
  omiga,
  s,
  epson = 0.01,
  lambda_L = 1e-05,
  lambda_0 = 0.1,
  sigma = 0.5,
  elta = 1e-06,
  alpha_min = 1e-15,
  alpha_max = 1e+15,
  alpha_min_newton = 1e-05,
  L = 10,
  p = 1,
  q = 5
)
}
\arguments{
\item{n}{the number of variables for a multinormal distribution}

\item{kkk}{the l0 constraint, i.e. a maximally allowable number of nonzeros}

\item{omiga}{a sparsity constraint set which includes the indexes of variables known to be conditionally independent}

\item{s}{the sample covariance}

\item{epson}{the first-order gradient of the objective function}

\item{lambda_L}{the last point of the decreased sequence}

\item{lambda_0}{the first point of the decreased sequence}

\item{sigma}{a parameter takes value in (0,1) to ensure the stepsize is sufficiently small}

\item{elta}{a positive parameter to ensure the objective function value is sufficiently decreased}

\item{alpha_min}{the minimum stepsize for Newton's method}

\item{alpha_max}{the maximum stepsize for Newton's method}

\item{alpha_min_newton}{the first-order gradient of the objective function}

\item{L}{length of a decreased sequence}

\item{p}{the remainder of the iteration k mod q}

\item{q}{every q − 1 iterations of the Newton steps we perform a single iteration of the gradient projection step}
}
\value{
the cholesky decompositon of the minimum point
}
\description{
a homotopy approximate Newton algorithm to compute MLE of the inverse covariance
}
\examples{
\dontrun{
library(MASS)
n = 100
A = AR2(n)
m = rep(0,n)
data <- mvrnorm(2*n,m,solve(A))
s = var(data)
omiga = which(A==0)
x = HANA(n,494,omiga,s)
t(x) \%*\% x
}
}
