% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{likelihood}
\alias{likelihood}
\title{The log-likelihood performance metric}
\usage{
likelihood(s, x)
}
\arguments{
\item{s}{the ground truth covariance}

\item{x}{the estimated inverse of the covariance}
}
\value{
d the log-likelihood performance metric
}
\description{
Compute the log-likelihood performance metric
}
\examples{
\dontrun{
s = diag(3)
x = matrix(c(1:9),3)
likelihood(s,x)
}
}
