% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{LossE}
\alias{LossE}
\title{The normalized entropy loss}
\usage{
LossE(s, x, n)
}
\arguments{
\item{s}{the ground truth covariance}

\item{x}{the estimated inverse of the covariance}

\item{n}{number of variables}
}
\value{
d the normalized entropy loss
}
\description{
Compute the normalized entropy loss
}
\examples{
\dontrun{
s = diag(3)
x = matrix(c(1:9),3)
LossE(s,x,3)
}
}
