% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{LossQ}
\alias{LossQ}
\title{The quadratic loss}
\usage{
LossQ(s, x, n)
}
\arguments{
\item{s}{the ground truth covariance}

\item{x}{the estimated inverse of the covariance}

\item{n}{number of variables}
}
\value{
d the quadratic loss
}
\description{
Compute the quadratic loss
}
\examples{
\dontrun{
s = diag(3)
x = matrix(c(1:9),3)
LossQ(s,x,3)
}
}
