% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{pp}
\alias{pp}
\title{A projection operator}
\usage{
pp(x, kap, omiga)
}
\arguments{
\item{x}{the operated matrix}

\item{kap}{the l0 constraint}

\item{omiga}{the space}
}
\value{
a projection of the operated matrix
}
\description{
Project a matrix into a space Omiga
}
\examples{
\dontrun{
x = matrix(c(1:9),3)
omiga = c(3,6,7,8)
pp(x,5,omiga)
}
}
