\name{StatComp22067-package}
\alias{StatComp22067-package}
\alias{StatComp22067}
\docType{package}
\title{
\packageTitle{StatComp22067}
}
\description{
\packageDescription{StatComp22067}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{StatComp22067}
\packageIndices{StatComp22067}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{StatComp22067}

Maintainer: \packageMaintainer{StatComp22067}
}
\references{
22067 <wangrc1960@mail.ustc.edu.cn>
}
\keyword{ package }

