% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{TPR}
\alias{TPR}
\title{True positive rate}
\usage{
TPR(s, x)
}
\arguments{
\item{s}{the ground truth inverse of the covariance}

\item{x}{the estimated inverse of the covariance}
}
\value{
k1/k2 true positive rate
}
\description{
Compute true positive rate of the estimation
}
\examples{
\dontrun{
s = diag(3)
x = matrix(c(1:9),3)
TPR(s,x)
}
}
